requirements.txt
asgiref==3.11.0
Django==4.2
django-multiselectfield==1.0.1
django-recaptcha==4.0.0
sqlparse==0.5.1

device-detector==5.0.1
geoip2==4.6.0
maxminddb==2.2.0

certifi==2025.4.26
charset-normalizer==2.0.12
idna==3.10
pytz==2025.2
PyYAML==6.0.1
regex==2023.8.8
requests==2.27.1
urllib3==1.26.20
typing_extensions==4.1.1